@@ -60,7 +60,7 @@
     "@typescript-eslint/parser": "^1.4.2",
     "chalk": "1.x.x",
     "devtron": "^1.4.0",
-    "electron": "4.1.0",
+    "electron": "4.2.2",
     "electron-installer-dmg": "0.2.x",
     "electron-packager": "13.0.x",
     "electron-winstaller": "2.x.x",
