@@ -107,7 +107,9 @@ module.exports = {
         },
         { label: localized('Star'), command: 'core:star-item' },
         { label: localized('Snooze') + '...', command: 'core:snooze-item' },
-        { label: localized('Share Link') + '...', command: 'core:share-item-link' },
+        { type: 'separator' },
+        { label: localized('Share this thread') + '...', command: 'core:share-item-link' },
+        { label: localized('Copy mailbox permalink'), command: 'core:copy-mailbox-link' },
         { type: 'separator' },
         { label: localized('Remove from view'), command: 'core:remove-from-view' },
         { label: localized('Remove and show next'), command: 'core:remove-and-next' },
